Added Email and JSON files from NOMIS maintainer
I was curious to see if there was a way to look up datasets by geography type using the API. Spencer, the maintainer confirmed that there was not at this point but helpfully provided some JSON files with the information in. Caveats that they are not checked and will never be updated apply.